Update byron_gh.py
Refactored the program. It's now much more structured, and a lot less brute force if/elif/else.

Mainly done because the first version was quick and dirty, for my eyes only. Now that it's squirmed its way into the public view, I decided that I needed to show that I actually do know how to program. :)